requirement.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
docutils==0.21.2
gcloud==0.18.3
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
jwcrypto==1.5.6
Kivy==2.3.0
Kivy-Garden==0.1.5
oauth2client==4.1.3
protobuf==5.29.0rc1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
pyparsing==3.2.0
Pyrebase4==4.8.0
python-jwt==4.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==75.2.0
six==1.16.0
typing_extensions==4.12.2
urllib3==1.26.20